+++ test/dummy/app/controllers/pages_controller.rb
@@ -882,6 +882,10 @@ class PagesController < ApplicationController
     render layout: "fullpage"
   end
 
+  def hero_centered_image_on_light
+    render layout: "fullpage"
+  end
+
   def hero_screenshot
     render layout: "fullpage"
   end